requirements.txt
kagglehub==0.3.10
keras==3.9.0
matplotlib==3.10.1
numpy==2.2.3
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
torch==2.6.0
torchinfo==1.8.0
torchvision==0.21.0
torchviz==0.0.3
tqdm==4.67.0

